Custom Markers and Quests
Custom markers now appear and disappear as intended. Only one can appear at a time, and if there is one that exists when the game starts up, it appears on the map and then can be removed as if it were placed.

Also added the functionality to remove quest icons from the compass.